requirements.txt
attrs==19.3.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colour==0.1.5
configparser==4.0.2
contextlib2==0.6.0.post1
cryptography==2.8
dingtalk==0.0.5
dingtalk-sdk==1.3.4
Django==1.11.11
djangorestframework==3.9.4
enum34==1.1.6
et-xmlfile==1.0.1
functools32==3.2.3.post2
huBarcode==1.0.0
idna==2.8
importlib-metadata==1.3.0
ipaddress==1.0.23
jdcal==1.4.1
jsonschema==3.2.0
more-itertools==5.0.0
numpy==1.16.5
openpyxl==2.6.4
optionaldict==0.1.1
pandas==0.21.1
pathlib2==2.3.5
Pillow==6.2.1
pyBarcode==0.8b1
pycparser==2.19
pygame==1.9.6
PyMySQL==0.9.3
pyrsistent==0.15.6
python-dateutil==2.8.1
pytz==2019.3
qrcode==6.1
requests==2.22.0
scandir==1.10.0
simplejson==3.17.0
six==1.13.0
StyleFrame==2.0.5
urllib3==1.25.7
xlrd==1.2.0
zipp==0.6.0
